refactor(aware_interfaces): extract all aware interfaces.
