+++ lib/angular.dart
@@ -5,6 +5,7 @@ export 'package:angular/application.dart';
 export 'package:angular/core/module.dart';
 export 'package:angular/directive/module.dart';
 export 'package:angular/core/annotation.dart';
+export 'package:angular/core/aware_interface.dart';
 export 'package:angular/tracing.dart';
 export 'package:angular/introspection.dart' hide
     elementExpando, publishToJavaScript;
