+++ lib/core/annotation.dart
@@ -3,13 +3,7 @@
  */
 library angular.core.annotation;
 
-import "dart:html" show ShadowRoot;
-
 export "package:angular/core/annotation_src.dart" show
-    AttachAware,
-    DetachAware,
-    ShadowRootAware,
-
     Formatter,
     DirectiveBinder,
     DirectiveBinderFn,
@@ -25,15 +19,4 @@ export "package:angular/core/annotation_src.dart" show
     NgOneWay,
     NgOneWayOneTime,
     NgTwoWay;
-
-
-/**
- * Implementing components [onShadowRoot] method will be called when
- * the template for the component has been loaded and inserted into Shadow DOM.
- * It is guaranteed that when [onShadowRoot] is invoked, that shadow DOM
- * has been loaded and is ready.
- */
-abstract class ShadowRootAware {
-  void onShadowRoot(ShadowRoot shadowRoot);
-}
