+++ lib/core_dom/module_internal.dart
@@ -15,6 +15,7 @@ import 'package:angular/utils.dart';
 import 'package:angular/cache/module.dart';
 
 import 'package:angular/core/annotation.dart';
+import 'package:angular/core/aware_interface.dart';
 import 'package:angular/core/module_internal.dart';
 import 'package:angular/core/parser/parser.dart';
 import 'package:angular/core_dom/dom_util.dart' as util;
