+++ test/angular_spec.dart
@@ -84,11 +84,12 @@ main() {
         "angular.app.Application",
         "angular.cache.CacheRegister",
         "angular.cache.CacheRegisterStats",
-        "angular.core.annotation.ShadowRootAware",
-        "angular.core.annotation_src.AttachAware",
+        "angular.core.aware_interface.AttachAware",
+        "angular.core.aware_interface.DetachAware",
+        "angular.core.aware_interface.ScopeAware",
+        "angular.core.aware_interface.ShadowRootAware",
         "angular.core.annotation_src.Component",
         "angular.core.annotation_src.Decorator",
-        "angular.core.annotation_src.DetachAware",
         "angular.core.annotation_src.Directive",
         "angular.core.annotation_src.DirectiveAnnotation",
         "angular.core.annotation_src.DirectiveBinder",
@@ -142,7 +143,6 @@ main() {
         "angular.core_internal.Interpolate",
         "angular.core_internal.RootScope",
         "angular.core_internal.Scope",
-        "angular.core_internal.ScopeAware",
         "angular.core_internal.ScopeDigestTTL",
         "angular.core_internal.ScopeEvent",
         "angular.core_internal.ScopeStats",
